// src/config/types.d.ts
export type ConfigTypes = {
  EMAIL_VERIFICATION_URL?: string;
  NODE_ENV: string;
  APP_NAME: string;
  PORT: number;
  LOCAL_PORT: number;
  APP_URL: string;
  MONGODB_URI: string;
  JWT_SECRET: string;
  SESSION_SECRET: string;
  REFRESH_TOKEN_SECRET: string;
  STRIPE_SECRET_KEY: string;
  STRIPE_WEBHOOK_SECRET: string;
  ACCESS_TOKEN_VALIDITY: number;
  REDIS_HOST: string;
  REDIS_PORT: number;
  REDIS_PASSWORD: string;
  REDIS_USERNAME: string;
  REDIS_DB: number;
  SUPPORT_EMAIL: string;
  COMMUNITY_EMAIL: string;
  };